+++ ego_metrology/__init__.py
@@ -1,7 +1,10 @@
+from .version import __version__
 from .profiler import EgoProfiler, ModelSectorConfig, ProfileResult, SECTOR_CONFIGS
 
-__version__ = "0.1.0"
-__author__  = "Julien Tournier"
-__license__ = "MIT"
-
-__all__ = ["EgoProfiler", "ModelSectorConfig", "ProfileResult", "SECTOR_CONFIGS"]
+__all__ = [
+    "__version__",
+    "EgoProfiler",
+    "ModelSectorConfig",
+    "ProfileResult",
+    "SECTOR_CONFIGS",
+]